単語詳細のトップページは削除
